Update README.md
The salesman problem is a problem in which, given the geographical coordinates of different locations in a city, 
one should find the shortest path in which all nodes have been traveled. The 'TSP_using_Genetic_Algorithm.ipynb' is a Python implementation of the Genetic Algorithm to solve the salesman problem for 7 different cities. Namely, New York, Cincinnati, Denver and etc. A detailed description of the assignment can be viewed via the pdf file and the final report of our study is detailed in the 'Project1_Team 4.docx' file.